+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   * Therion run output should be selectable as a whole and not only single lines.
   * Therion run output has no line breaks after copy/paste.
   * Searching for Therion install path on windows always failing.
+  * Therion run on Windows.
 * Infrastructure maintenance:
   * Included "Current status" section in README.md. [requested by CaverBruce]
   * Included "Choose THConfig and run Therion" and "Run Therion" buttons in both MapiaHome and TH2FileEdit pages.